Move platform shortcuts to backend
