+++ src-backend/src/lib.rs
@@ -9,3 +9,4 @@ pub mod build;
 pub mod hex_format;
 pub mod regions;
 pub mod midi;
+pub mod shortcuts;
