+++ src/utils/window.ts
@@ -3,19 +3,19 @@ import { hasActionKey } from './query/shortcut-key'
 import { invoke } from '@tauri-apps/api'
 
 export function setupWindow() {
-  window.addEventListener('keydown', (event) => {
-    // Prevent the Ctrl + R refresh.
-    if (hasActionKey(event) && event.key === 'r') {
+  if (window.__TAURI__) {
+    window.addEventListener('keydown', (event) => {
+      // Prevent the Ctrl + R refresh.
+      if (hasActionKey(event) && event.key === 'r') {
+        event.preventDefault()
+      }
+    })
+
+    const handler = (event: Event) => {
+      // Don't bring up the "reload" context menu. It's not great!
       event.preventDefault()
     }
-  })
 
-  const handler = (event: Event) => {
-    // Don't bring up the "reload" context menu. It's not great!
-    event.preventDefault()
-  }
-
-  if (window.__TAURI__) {
     invoke('is_debug').then(debug => {
       if (!debug) {
         window.addEventListener('contextmenu', handler)
